--- src/modules/offers.js
@@ -1,5 +1,0 @@
-export const prepareOfferString = (offer) => `
-    <li>
-      <button class="trip-point__offer">${offer.label} +&euro;&nbsp;${offer.cost}</button>
-    </li>
-`;
