+++ src/modules/sorting.js
@@ -1,27 +1,10 @@
-import {createElement} from './common/utils';
+import {Component} from './common/component';
 
 const columns = [`event`, `time`, `price`];
 
-export class Sorting {
+export class Sorting extends Component {
   constructor() {
-    this._element = null;
-  }
-
-  create() {
-    if (this._element) {
-      this.destroy();
-    }
-
-    this._element = createElement(this.template);
-    return this._element;
-  }
-
-  destroy() {
-    this._element = null;
-  }
-
-  get element() {
-    return this._element;
+    super();
   }
 
   get template() {
